Abstract DB connection functions
